Moving RequestLogFactory back to its original package to minimise diffs in PR
