+++ dropwizard-core/src/main/java/io/dropwizard/server/AbstractServerFactory.java
@@ -18,7 +18,7 @@ import io.dropwizard.jersey.setup.JerseyEnvironment;
 import io.dropwizard.jetty.GzipFilterFactory;
 import io.dropwizard.jetty.MutableServletContextHandler;
 import io.dropwizard.jetty.NonblockingServletHolder;
-import io.dropwizard.jetty.logging.RequestLogFactory;
+import io.dropwizard.jetty.RequestLogFactory;
 import io.dropwizard.lifecycle.setup.LifecycleEnvironment;
 import io.dropwizard.servlets.ThreadNameFilter;
 import io.dropwizard.util.Duration;
